:truck: rename(core): core 패키지 안으로 config 패키지 이동
