+++ src/main/java/ac/dnd/dodal/common/constant/Constants.java
@@ -21,6 +21,15 @@ public class Constants {
             //Swagger URL
             //로그인 없이 둘러볼 수 있는 URL
 
+            // 회원가입
+            "/api/auth/sign-up",
+            // 로그인
+            "/api/auth/sign-in",
+            "/api/auth/login/**",
+            "/api/auth/login/kakao",
+            "/api/auth/login/naver",
+            "/api/auth/login/google",
+            "/api/auth/login/apple"
     );
 
 }
