+++ src/main/java/ac/dnd/dodal/core/config/security/filter/CustomLogoutFilter.java
@@ -1,6 +1,6 @@
-package ac.dnd.dodal.config.security.filter;
+package ac.dnd.dodal.core.config.security.filter;
 
-import ac.dnd.dodal.config.security.enums.E_security_code;
+import ac.dnd.dodal.core.config.security.enums.E_security_code;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -25,9 +25,12 @@ public class CustomLogoutFilter extends OncePerRequestFilter {
         if ("/api/auth/sign-out".equals(httpRequest.getRequestURI()) && "POST".equalsIgnoreCase(
                 httpRequest.getMethod())) {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            System.out.println("🔴 인증 실패: SecurityContext에 인증 정보 없음");
 
             // 인증된 사용자가 아니라면 401 응답을 반환
             if (authentication == null || !authentication.isAuthenticated()) {
+                System.out.println("🔴 인증 실패: SecurityContext에 인증 정보 없음");
+
                 httpResponse.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
                 httpResponse.setContentType("application/json;charset=UTF-8");
 
@@ -37,6 +40,8 @@ public class CustomLogoutFilter extends OncePerRequestFilter {
                 responseData.put("error", E_security_code.FAILURE_LOGOUT.getMessage());
 
                 httpResponse.getWriter().write(JSONValue.toJSONString(responseData));
+
+                httpResponse.flushBuffer();
                 return; // 필터 체인을 끊고 응답을 반환
             }
         }
